pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "fba_net"
version = "0.1.0"
authors = [{ name = "Yujing Sun", email = "yujingsun1999@gmail.com" }]
requires-python = ">=3.11"
license = { text = "MIT" }
description = "Towards Real-World Burst Image Super-Resolution: Benchmark and Method"
dependencies = [
    "jaxlib[cuda12_pip]@https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.18+cuda12.cudnn89-cp311-cp311-manylinux2014_x86_64.whl",
    "jax[cuda12_pip]@https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.18+cuda12.cudnn89-cp311-cp311-manylinux2014_x86_64.whl",
    "jaxtyping@https://github.com/google/jaxtyping/archive/7a84b27da9e57c425ce4e6333121c3cdf2e07302.tar.gz",
    "equinox@https://github.com/patrick-kidger/equinox/archive/5d8a06e5e931295d9fef70f240a33a7061a1b350.tar.gz",
    "optax@https://github.com/google-deepmind/optax/archive/9f2f5caeca3030accf608cce934182798b5da7c6.tar.gz",
    "dm-pix@https://github.com/google-deepmind/dm_pix/archive/14ec005685c4933420939010e774afde20c9140a.tar.gz",
    "einops@https://github.com/arogozhnikov/einops/archive/a6e93530ec2dce44f473e6065fad4d8236cda4f3.tar.gz",
    # For DALI
    "nvidia-dali-cuda120@https://developer.download.nvidia.com/compute/redist/nvidia-dali-cuda120/nvidia_dali_cuda120-1.30.0-9783408-py3-none-manylinux2014_x86_64.whl",
    "nvidia-nvjpeg-cu12",
    "nvidia-npp-cu12",
    "nvidia-cuda-runtime-cu12",
    "Pillow"
]

[project.optional-dependencies]
format = ["black", "ruff"]
type = ["beartype", "typing-extensions"]

[tool.black]
line-length = 120
target-version = ['py312']

[tool.ruff]
preview = true
select = ["F", "E", "W", "I", "PL", "FURB", "LOG", "RUF", "G", "TID"]
ignore = [
    # Ignore functions with many arguments (I like currying)
    "PLR0913",
    # Ignore syntax error in forward annotation for jaxtyping
    "F722",
]
line-length = 120

[tool.pyright]
include = ["./fba_net"]
pythonVersion = "3.11"
pythonPlatform = "Linux"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
